@@ -1,6 +1,7 @@
 {
   "name": "owlbear-rodeo",
   "version": "1.10.2",
+  "type": "module",
   "private": true,
   "dependencies": {
     "@babylonjs/core": "4.2.1",
@@ -9,13 +10,13 @@
     "@dnd-kit/sortable": "^4.0.0",
     "@mitchemmc/dexie-export-import": "^1.0.1",
     "@msgpack/msgpack": "^2.8.0",
-    "@react-spring/konva": "9.4.4",
+    "@react-spring/konva": "9.7.4",
     "@tensorflow/tfjs": "^3.15.0",
-    "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/react": "^13.0.0",
     "@testing-library/user-event": "^14.4.3",
+    "@vitejs/plugin-react": "^4.3.1",
     "ajv": "^8.11.2",
-    "ammo.js": "kripken/ammo.js#85c0614cd5338aa0843814fe7bbd2df48164ece7",
+    "ammojs-typed": "^1.0.6",
     "case": "^1.6.3",
     "color": "^3.2.1",
     "comlink": "^4.3.1",
@@ -39,22 +40,21 @@
     "polygon-clipping": "^0.15.3",
     "pretty-bytes": "^6.0.0",
     "raw.macro": "^0.4.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-intersection-observer": "8.33.1",
     "react-konva": "^17.0.2-6",
-    "react-konva-utils": "^0.2.0",
-    "react-markdown": "4",
-    "react-media": "^2.0.0-rc.1",
+    "react-konva-utils": "^1.0.0",
+    "react-markdown": "^8.0.0",
     "react-modal": "^3.14.4",
-    "react-resize-detector": "^7.0.0",
+    "react-resize-detector": "^9.0.0",
+    "react-responsive": "^10.0.0",
     "react-router-dom": "^5.1.2",
     "react-router-hash-link": "^2.4.3",
-    "react-scripts": "^4.0.3",
     "react-select": "^5.2.2",
     "react-spring": "9.4.4",
     "react-textarea-autosize": "^8.3.3",
-    "react-toast-notifications": "^2.5.1",
+    "react-toastify": "^10.0.5",
     "react-use-gesture": "^9.1.3",
     "shortid": "^2.2.15",
     "simple-peer": "^9.11.1",
@@ -63,47 +63,34 @@
     "socket.io-client": "^4.5.4",
     "socket.io-msgpack-parser": "^3.0.2",
     "source-map-explorer": "^2.5.3",
-    "theme-ui": "^0.10.0",
+    "theme-ui": "^0.16.2",
     "tiny-typed-emitter": "^2.1.0",
-    "use-image": "1.0.10",
+    "use-image": "^1.1.1",
     "uuid": "^8.3.2",
+    "vite": "^5.4.6",
+    "vite-tsconfig-paths": "^5.0.1",
     "webrtc-adapter": "^8.2.0"
   },
   "scripts": {
-    "analyze": "source-map-explorer 'build/static/js/*.js'",
-    "start": "react-scripts start",
-    "build": "react-scripts --max_old_space_size=4096 build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
   },
   "eslintConfig": {
     "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
   },
   "devDependencies": {
     "@types/color": "^3.0.3",
     "@types/deep-diff": "^1.0.0",
     "@types/file-saver": "^2.0.5",
-    "@types/jest": "^27.4.1",
     "@types/lodash.chunk": "^4.2.7",
     "@types/lodash.clonedeep": "^4.5.7",
     "@types/lodash.get": "^4.4.7",
     "@types/lodash.set": "^4.3.7",
-    "@types/node": "^15.6.0",
+    "@types/node": "^18.0.0",
     "@types/normalize-wheel": "^1.0.1",
-    "@types/react": "^17.0.6",
-    "@types/react-dom": "^17.0.5",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@types/react-modal": "^3.13.1",
     "@types/react-router-dom": "^5.1.7",
     "@types/react-select": "^5.0.1",
@@ -111,6 +98,8 @@
     "@types/simple-peer": "^9.11.5",
     "@types/uuid": "^8.3.4",
     "typescript": "^4.2.4",
+    "vite-plugin-babel-macros": "^1.0.6",
+    "vite-plugin-node-polyfills": "^0.22.0",
     "worker-loader": "^3.0.8"
   }
 }
