@@ -1,4 +1,4 @@
-import ReactDOM from "react-dom";
+import { createRoot } from 'react-dom/client';
 import App from "./App";
 import Modal from "react-modal";
 
@@ -21,7 +21,9 @@ if (!("IntersectionObserver" in window)) {
 
 Modal.setAppElement("#root");
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const container = document.getElementById("root")
+const root = createRoot(container!);
+root.render(<App />)
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
