@@ -1,4 +1,4 @@
-import { ThemeProvider } from "theme-ui";
+import { ThemeUIProvider } from "theme-ui";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
 import theme from "./theme";
@@ -14,47 +14,59 @@ import { SettingsProvider } from "./contexts/SettingsContext";
 import { KeyboardProvider } from "./contexts/KeyboardContext";
 import { DatabaseProvider } from "./contexts/DatabaseContext";
 import { UserIdProvider } from "./contexts/UserIdContext";
+import { Bounce, ToastContainer } from "react-toastify";
+import 'react-toastify/dist/ReactToastify.css';
 
-import { ToastProvider } from "./components/Toast";
 
 function App() {
   return (
-    <ThemeProvider theme={theme}>
+    <ThemeUIProvider theme={theme}>
       <SettingsProvider>
         <AuthProvider>
           <KeyboardProvider>
-            <ToastProvider>
-              <Router>
-                <Switch>
-                  <Route path="/how-to">
-                    <HowTo />
-                  </Route>
-                  <Route path="/release-notes">
-                    <ReleaseNotes />
-                  </Route>
-                  <Route path="/about">
-                    <About />
-                  </Route>
-                  <Route path="/faq">
-                    <FAQ />
-                  </Route>
-                  <Route path="/game/:id">
-                    <DatabaseProvider>
-                      <UserIdProvider>
-                        <Game />
-                      </UserIdProvider>
-                    </DatabaseProvider>
-                  </Route>
-                  <Route path="/">
-                    <Home />
-                  </Route>
-                </Switch>
-              </Router>
-            </ToastProvider>
+            <ToastContainer
+              position="top-right"
+              autoClose={5000}
+              hideProgressBar={false}
+              newestOnTop={false}
+              closeOnClick
+              rtl={false}
+              pauseOnFocusLoss
+              draggable
+              pauseOnHover
+              theme="light"
+              transition={Bounce}
+            />
+            <Router>
+              <Switch>
+                <Route path="/how-to">
+                  <HowTo />
+                </Route>
+                <Route path="/release-notes">
+                  <ReleaseNotes />
+                </Route>
+                <Route path="/about">
+                  <About />
+                </Route>
+                <Route path="/faq">
+                  <FAQ />
+                </Route>
+                <Route path="/game/:id">
+                  <DatabaseProvider>
+                    <UserIdProvider>
+                      <Game />
+                    </UserIdProvider>
+                  </DatabaseProvider>
+                </Route>
+                <Route path="/">
+                  <Home />
+                </Route>
+              </Switch>
+            </Router>
           </KeyboardProvider>
         </AuthProvider>
       </SettingsProvider>
-    </ThemeProvider>
+    </ThemeUIProvider>
   );
 }
 
